Link analysis to sourceMeter stats

+++ jthemis/jthemis-core/src/main/java/net/sourceforge/joceanus/jthemis/sourcemeter/ThemisSMClass.java
@@ -50,7 +50,7 @@ public class ThemisSMClass
     /**
      * The stats.
      */
-    private final Map<ThemisSMStat, Long> theStats;
+    private final Map<ThemisSMStat, Integer> theStats;
 
     /**
      * The parent.
@@ -91,7 +91,7 @@ public class ThemisSMClass
     }
 
     @Override
-    public Map<ThemisSMStat, Long> getStatistics() {
+    public Map<ThemisSMStat, Integer> getStatistics() {
         return theStats;
     }
 
@@ -121,7 +121,7 @@ public class ThemisSMClass
 
     @Override
     public void setStatistic(final ThemisSMStat pStat,
-                             final Long pValue) {
+                             final Integer pValue) {
         theStats.put(pStat, pValue);
     }
